requirements.txt
##Web Framework and Server Dependencies
fastapi==0.111.0
uvicorn==0.29.0
starlette==0.37.2

##library for secure session management (cookies)
itsdangerous==2.2.0

##LangChain and LangGraph Core
langchain==0.3.26
langgraph==0.5.4
langchain-core==0.3.70
langchain-google-genai==2.1.8
langchain-community==0.3.27
langchain-tavily==0.2.10
langchain-text-splitters==0.3.8

##Google API Libraries
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

##Agent Tools
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
requests==2.32.4

##Utilities
python-dotenv==1.1.1
python-dateutil==2.9.0

##Core Dependencies (resolved from installation)
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
dataclasses-json==0.6.7
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
ormsgpack==1.10.0
orjson==3.11.0
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
PyYAML==6.0.2
pyparsing==3.2.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-extensions==4.14.1
typing-inspect==0.9.0
uritemplate==4.2.0
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0